Move components to a separate folder and file
